Start implementation of a PropertyHelper (focus on titles first)
The purpose of a property helper is to provide useful methods
to handle properties and ressources with certain properties.

Think about titles: A title helper is helpful in cases where you want
the title of all your ressources.

This implementation provides a basic index, which caches title
information for each related resource. You need to create an
index first, using a key-value-pair based caching solution. After
you created it, you can ask the index to give you titles for
different lists of URIs.

The way this PropertyHelper is implemented helps you to integrate
it in different scenarios. Our main approach would be to use it
inside an endpoint, which receives an URI list and responses with
a list of titles, if available.